+++ src/components/lumalike/template1.tsx
@@ -10,7 +10,7 @@ import { EventDescription } from './template1-components/event-description'
 import { EventLocation } from './template1-components/event-location'
 import MainLandingNav from '../mainLanding/mainNav'
 import { useTranslations } from 'next-intl'
-import { ManageMatchmaking } from '../events/ ManageMatchmaking'
+import { ManageMatchmaking } from '../events/ManageMatchmaking'
 import { PublicEventType, CustomFieldType } from '@/schemas/eventSchema'
 import { LuciaUserType } from '@/schemas/userSchema'
 import { NetworkingMatchesType } from '@/schemas/networkingMatchSchema'
